requirements.txt
# Основные зависимости
ragbits>=1.0.0
python-telegram-bot>=20.0
pandas>=1.5.0
httpx>=0.24.0

# LangChain для загрузки документов
langchain>=0.1.0
langchain-community>=0.0.10

# Embedding модели
sentence-transformers>=2.2.0

# Дополнительные утилиты  
python-dotenv>=1.0.0
asyncio-mqtt>=0.13.0
